Change: don't error log for 404s
Changes what is logged when bib id doesn't match a bib to issue a INFO
level log rather than an ERROR level log (which was causing lots of
false positives in our alarms). Includes tests asserting the new policy.

Also cleans up README to clarify installation and contributing flow. Also
cleans up swagger document to add several missing params, retire unused
swagger files, and rename it to better reflect its relationship with the
app version.
