+++ config/production.env
@@ -8,3 +8,4 @@ NYPL_OAUTH_SECRET=AQECAHh7ea2tyZ6phZgT4B9BDKwguhlFtRC6hgt+7HbmeFsrsgAAAIcwgYQGCS
 NYPL_API_BASE_URL=https://platform.nypl.org/api/v0.1/
 NYPL_CORE_VERSION=v1.31
 PORT=8081
+SEARCH_ITEMS_SIZE=100
