+++ config/test.env
@@ -1,0 +1,6 @@
+# Number of items returned for each bib by default:
+SEARCH_ITEMS_SIZE=100
+
+# This can be any port, but let's choose something reasonably high to
+# avoid conflict:
+PORT=5678
