+++ lib/resources.js
@@ -206,9 +206,12 @@ module.exports = function (app, _private = null) {
         resp.hits.hits[0]._source.items = resp.hits.hits[0].inner_hits.items.hits.hits.map((item) => item._source)
       }
       // Mindfully throw errors for known issues:
-      if (!resp || !resp.hits) throw new Error('Error connecting to index')
-      else if (resp && resp.hits && resp.hits.total === 0) throw new errors.NotFoundError('Record not found')
-      else {
+      if (!resp || !resp.hits) {
+        throw new Error('Error connecting to index')
+      } else if (resp && resp.hits && resp.hits.total === 0) {
+        app.logger.info(`Record not found: ${params.uri}`)
+        throw new errors.NotFoundError(`Record not found: ${params.uri}`)
+      } else {
         let massagedResponse = new ResponseMassager(resp)
         return massagedResponse.massagedResponse(request)
           .catch((e) => {
