+++ package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "chai": "^4.1.2",
+    "chai-as-promised": "^7.1.1",
     "md5": "^2.2.1",
     "minimist": "^1.2.5",
     "mocha": "^6.2.3",
