+++ test/fixtures/query-aa79df993fc122416c62a0509aa6fdb1.json
@@ -1,0 +1,14 @@
+{
+  "took": 1,
+  "timed_out": false,
+  "_shards": {
+    "total": 3,
+    "successful": 3,
+    "failed": 0
+  },
+  "hits": {
+    "total": 0,
+    "max_score": null,
+    "hits": []
+  }
+}
