+++ test/fixtures/query-cc69301081741e1789e00bc5bee60461.json
@@ -1,0 +1,9 @@
+{
+  "took": 2,
+  "timed_out": false,
+  "_shards": {
+    "total": 3,
+    "successful": 3,
+    "failed": 0
+  }
+}
