+++ test/test_helper.js
@@ -1,14 +1,11 @@
+const chai = require('chai')
+const chaiAsPromised = require('chai-as-promised')
+
 // Set some env variables:
-process.env.SEARCH_ITEMS_SIZE = 100
-// This can be any port, but let's choose something reasonably high to
-// avoid conflict:
-process.env.PORT = 5678
+require('dotenv').config({ path: './config/test.env' })
 
 // Establish base url for local queries:
-global.TEST_BASE_URL = 'http://localhost:' + process.env.PORT
-
-// By virtue of including app.js, we start listening on above port:
-require('../app.js')
+global.TEST_BASE_URL = `http://localhost:${process.env.PORT}`
 
 // Nullify SCSB creds just in case they've been brought in by app.js by a
 // local .env (but only if we're not updating fixtures, for which we'll need
@@ -19,4 +16,6 @@ if (!process.env.UPDATE_FIXTURES) {
 }
 
 require('../lib/globals')
-global.expect = require('chai').expect
+
+chai.use(chaiAsPromised)
+global.expect = chai.expect
